requirements.txt
streamlit
scikit-learn
pandas
numpy
joblib
pdfminer.six